+++ R/gradient.R
@@ -3,7 +3,9 @@
 #' Computes numerical derivatives, gradients and Jacobians. Supports both two-sided
 #' (central) and one-sided (forward or backward) derivatives. Calculations can be
 #' executed on multiple cores to cut down execution time for slow functions or
-#' to attain higher accuracy faster.
+#' to attain higher accuracy faster. Currently for Mac & Linux only because
+#' Windows cannot handle \code{parallel::mclapply()}. A \code{parallel::parLapply}
+#' version is in development.
 #'
 #' @param FUN A function returning a numeric scalar or a vector.
 #'   If the function returns a vector, the output will be is a Jacobian.
@@ -11,40 +13,37 @@
 #'   it will be reassigned to \code{FUN} with a warning.
 #' @param x Numeric vector or scalar: point at which the derivative is estimated.
 #'   \code{FUN(x)} must return a finite value.
-#' @param h Numeric scalar or vector specifying the step size for the numerical
-#'   difference. Must be length 1 or match \code{length(x)}.
-#'   If too large, the slope of the secant poorly estimates the derivative;
-#'   if too small, it leads to numerical instability due to the function value rounding.
+#' @param h Numeric scalar, vector, or character specifying the step size for the numerical
+#'   difference. If character (\code{"CR"}, \code{"CRm"}, \code{"DV"}, or \code{"SW"}),
+#'   calls \code{gradstep()} with the appropriate step-selection method.
+#'   Must be length 1 or match \code{length(x)}.
 #' @param deriv.order Integer indicating the derivative order,
 #'   \eqn{\mathrm{d}^m / \mathrm{d}x^m}{d^m/dx^m}.
 #' @param acc.order Integer specifying the desired accuracy order.
 #'   The error typically scales as \eqn{O(h^{\mathrm{acc.order}})}{O(h^acc.order)}.
 #' @param side Integer scalar or vector indicating difference type:
 #'   \code{0} for central, \code{1} for forward, and \code{-1} for backward differences.
-#'   Using \code{2} (for 'two-sided') triggers a warning and is treated as \code{0}.
-#'   with a warning.
 #'   Central differences are recommended unless computational cost is prohibitive.
 #' @param f0 Optional numeric scalar or vector: if provided and applicable, used
 #'   where the stencil contains zero (i.e. \code{FUN(x)} is part of the sum)
 #'   to save time. Currently ignored.
-#' @param parallel String indicating the parallel execution mode.
-#'   This argument is ignored if a valid \code{cluster} is provided,
-#'   as the cluster settings take precedence.
-#'   On Mac and Linux, use \code{"fork"} to invoke \code{mclapply} or \code{"PSOCK"}
-#'   to create a PSOCK cluster dynamically.
-#'   On Windows, use \code{"PSOCK"} if you have not created the cluster yet, although
-#'   using a dedicated cluster via \code{cluster} is highly recommended to reduce overhead.
-#'   \code{"auto"} switches to \code{"PSOCK"} on Windows (with a warning)
-#'   and to \code{"fork"} on everything else.
-#' @param cores Integer specifying the number of parallel processes to use.
-#' @param cluster An optional cluster object. Ensure that necessary objects and packages
-#'   are available in the cluster environment with \code{clusterExport()} and \code{clusterEvalQ()}.
+#' @param h0 Numeric scalar of vector: initial step size for automatic search with
+#'   \code{gradstep()}.
+#' @param method.args A named list of tuning parameters passed to \code{gradstep()}.
+#' @param cores Integer specifying the number of parallel processes to use. Recommended
+#'   value: the number of physical cores on the machine minus one.
 #' @param load.balance Logical: if \code{TRUE}, disables pre-scheduling for \code{mclapply()}
 #'   or enables load balancing with \code{parLapplyLB()}.
 #' @param func Deprecated; for \code{numDeriv::grad()} compatibility only.
+#' @param report Integer: if \code{0}, returns a gradient without any attributes; if \code{1},
+#'   attaches the step size and its selection method: \code{2} or higher, attaches the full
+#'   diagnostic output (overrides \code{diagnostics = FALSE} in \code{method.args}).
 #' @param ... Additional arguments passed to \code{FUN}.
 #'
 #' @details
+#'
+#' If the step size is too large, the slope of the secant poorly estimates the derivative;
+#' if it is too small, it leads to numerical instability due to the function value rounding.
 #'
 #' The optimal step size for one-sided differences typically approaches Mach.eps^(1/2)
 #' to balance the Taylor series truncation error with the rounding error due to storing
@@ -62,7 +61,18 @@
 #'   Unlike the output of \code{numDeriv::grad} and \code{numDeriv::jacobian},
 #'   this output preserves the names of \code{x} and \code{FUN(x)}.
 #'
+#' @seealso [gradstep()] for automatic step-size selection.
+#'
 #' @examples
+#'
+#' f <- function(x) sum(sin(x))
+#' g1 <- Grad(x = 1:4, FUN = f)
+#' g2 <- Grad(x = 1:4, FUN = f, h = 7e-6)
+#' g2 - g1 # Tiny differences due to different step sizes
+#' g3 <- Grad(x = 1:4, FUN = f, h = "SW")
+#' g3.full <- Grad(x = 1:4, FUN = f, h = "SW", report = 2)
+#' attr(g3.full, "step.search")$exitcode  # Success
+#'
 #' \dontrun{
 #' slowFun <- function(x) {Sys.sleep(0.05); print(x, digits = 12); sum(sin(x))}
 #' slowFunVec <- function(x) {Sys.sleep(0.05); print(x, digits = 12)
@@ -92,10 +102,10 @@
 Grad <- function(FUN, x,
                  deriv.order = 1L, side = 0,
                  acc.order = ifelse(abs(side) == 1, 1L, 2L),
-                 h = .Machine$double.eps^(1 / (deriv.order + acc.order)),
-                 f0 = NULL,
-                 parallel = c("auto", "fork", "PSOCK", "none"), cores = 2,
-                 cluster = NULL, load.balance = TRUE, func = NULL,
+                 h = (abs(x) + (x==0)) * .Machine$double.eps^(1 / (deriv.order + acc.order)),
+                 h0 = NULL, method.args = list(), f0 = NULL,
+                 cores = 2, load.balance = TRUE, func = NULL,
+                 report = 1L,
                  ...) {
   n <- length(x)
   if (!(length(side) %in% c(1, n))) stop("The 'side' argument must have length 1 or same length as x.")
@@ -106,7 +116,7 @@ Grad <- function(FUN, x,
       FUN <- func
       warning("Use the argument 'FUN' to pass the function for differencing to Grad instead of 'func'.")
     } else {
-      stop("Pass the function for differencing as the named argument 'FUN' or the first unnamed argument.")
+      stop("Pass the function for differencing as the named argument 'FUN' or as the first unnamed argument.")
     }
   }
 
@@ -114,10 +124,6 @@ Grad <- function(FUN, x,
   if (is.null(side)) side <- numeric(n) # NULL --> default central, 0
   if (length(side) == 1) side <- rep(side, n)
   side[!is.finite(side)] <- 0 # NA --> default 'central
-  if (any(is2 <- side == 2)) {
-    side[is2] <- 0 # Interpreting '2' as two-sided = central
-    warning("Interpreting 'side = 2' as '2-sided central differences'; please use side = 0.")
-  }
   if (!all(side %in% -1:1)) stop("The 'side' argument must contain values 0 for central, 1 for forward and 2 for backward difference.")
 
   ell <- list(...)
@@ -135,18 +141,26 @@ Grad <- function(FUN, x,
   # END compatibility with numDeriv::grad
   #######################################
 
+  if (!is.function(FUN)) stop("'FUN' must be a function.")
   if (length(deriv.order) == 1) deriv.order <- rep(deriv.order, n)
   if (length(acc.order) == 1) acc.order <- rep(acc.order, n)
   # side is already a vector
   # TODO: the part where step is compared to step.CR, step.DV etc.
+  autostep <- FALSE
+  if (is.character(h)) {
+    method <- h
+    if (report == 2) method.args$diagnostics <- TRUE
+    h.auto <- gradstep(x = x, FUN = FUN, h0 = h0, method = method, method.args = method.args, ...)
+    h <- h.auto$par
+    autostep <- TRUE
+    # TODO: use this gradient already
+  } else if (any(h <= 0)) {
+    stop("The argument 'h' (step size) must be positive.")
+  }
   if (length(h) == 1) h <- rep(h, n)
   if (length(deriv.order) != n) stop("The argument 'deriv.order' must have length 1 or length(x).")
   if (length(acc.order) != n) stop("The argument 'acc.order' must have length 1 or length(x).")
   if (length(h) != n) stop("The argument 'h' (step size) must have length 1 or length(x).")
-
-  parallel <- parallel[1]
-  parallel <- parallel != "none"
-  # TODO: proper clusters
 
   xlist <- lapply(1:n, function(i) {
     sw <- fdCoef(deriv.order = deriv.order[i], acc.order = acc.order[i], side = side[i])
@@ -167,14 +181,13 @@ Grad <- function(FUN, x,
   xvals <- lapply(seq_len(nrow(xdf)), function(i) xm[i, ])
 
   # Parallelising the task in the most efficient way possible, over all values of all grids
-  fvals <- if (parallel && cores > 1) {
+  fvals <- if (cores > 1) {
     parallel::mclapply(X = xvals, FUN = function(x) FUN(x, ...), mc.cores = cores, mc.preschedule = !load.balance)
   } else {
     lapply(xvals, function(x) FUN(x, ...))
   }
-  # else if (parallel & !use.mclapply & !is.null(cluster)) {
-  #if (load.balance) parallel::parLapplyLB(cl = cluster, X = xx, fun = function(x) func(x, ...)) else
-  #    parallel::parLapply(cl = cluster, X = xx, fun = function(x) func(x, ...))
+  if (any(!sapply(fvals, function(x) is.numeric(x) | is.na(x))))
+    stop("'FUN' must output numeric values only, but non-numeric values were returned.")
 
   # The output can be vector-valued, which is why we ensure that dimensions are not dropped
   fvals <- do.call(rbind, fvals)
@@ -190,5 +203,19 @@ Grad <- function(FUN, x,
     if (!is.null(names(x))) colnames(jac) <- names(x)
     if (!is.null(colnames(fvals))) rownames(jac) <- colnames(fvals)
   }
+
+  if (report > 0) {
+    attr(jac, "step.size") <- h
+    if (autostep) {
+      attr(jac, "step.size.method") <- method
+    } else if (all(h == (abs(x) + (x==0)) * .Machine$double.eps^(1 / (deriv.order + acc.order)))) {
+      attr(jac, "step.size.method") <- "default"
+    } else {
+      attr(jac, "step.size.method") <- "user-supplied"
+    }
+    if (autostep & report > 1) attr(jac, "step.search") <- h.auto
+  }
+
   return(jac)
 }
+