+++ R/zzz.R
@@ -1,6 +1,6 @@
 #' @importFrom Rdpack reprompt
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Parallel numerical derivatives v. 0.0.3 (2023-06-01).")
+  packageStartupMessage("Parallel numerical derivatives v. 0.0.4 (2024-06-10).")
   packageStartupMessage("This is *NOT* a stable version. Core functions subject to change.")
 }